Move FreeRTOS-Source content up a level
