+++ iris-fsw-softconsole/src/drivers/leds.c
@@ -21,8 +21,8 @@
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 /* Scheduler includes. */
-#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
-#include <FreeRTOS/FreeRTOS-Source/include/task.h>
+#include <FreeRTOS/include/FreeRTOS.h>
+#include <FreeRTOS/include/task.h>
 
 /* Library includes. */
 #include "drivers/mss_gpio/mss_gpio.h"
