+++ iris-fsw-softconsole/src/main.c
@@ -47,10 +47,10 @@
 #include "csp/interfaces/csp_if_can.h"
 
 /* Kernel includes. */
-#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
-#include <FreeRTOS/FreeRTOS-Source/include/queue.h>
-#include <FreeRTOS/FreeRTOS-Source/include/semphr.h>
-#include <FreeRTOS/FreeRTOS-Source/include/task.h>
+#include <FreeRTOS/include/FreeRTOS.h>
+#include <FreeRTOS/include/queue.h>
+#include <FreeRTOS/include/semphr.h>
+#include <FreeRTOS/include/task.h>
 
 /* Driver includes. */
 #include "drivers/mss_uart/mss_uart.h"    // For baud rate defines and instances
