+++ iris-fsw-softconsole/src/tasks/scheduler.c
@@ -16,8 +16,8 @@
 #include <string.h>
 
 // FreeRTOS Libraries
-#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
-#include <FreeRTOS/FreeRTOS-Source/include/task.h>
+#include <FreeRTOS/include/FreeRTOS.h>
+#include <FreeRTOS/include/task.h>
 
 // RTC Libraries
 #include "drivers/device/rtc/rtc_common.h"
