+++ iris-fsw-softconsole/tests/filesystem_tests.c
@@ -8,8 +8,8 @@
 // - Created.
 //-------------------------------------------------------------------------------------------------
 
-#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
-#include <FreeRTOS/FreeRTOS-Source/include/task.h>
+#include <FreeRTOS/include/FreeRTOS.h>
+#include <FreeRTOS/include/task.h>
 #include "tests.h"
 
 #include "drivers/filesystem_driver.h"
